create_configure_exe: eat the error from configure -help
Ignore error code generated by "configure -help" -command. Configure
is run with -help just to get the .exe, no need to do the
configuration. Running configure with -help makes configure exit
with error code, which then causes the create_configure_exe to fail.
If configure behaviour is changed later, this can be reverted, but for
now we need this to keep CI running.

Change-Id: I4d4976cfa8dd69ce8679370fd2f840eb989a104c
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>
Reviewed-by: Iikka Eklund <iikka.eklund@digia.com>