Helper script for online repository updates
This script can be used to update online repositories generated with
Qt Installer-Framework. The actual purpose for this tool is to ease out
online updates in mirrored production servers where you have no control
in which order files get mirrored. For Qt Installer-Framework we must
ensure the data files (*.7z) are in sync first before meta files
(Updates.xml) can be put in place as the meta files refer to the data files.
The script allows the user to perform updates in phases. In first phase the
user can update only the data files. In the second phase the user can update
the meta files for those components that had the data files updated previously.
The script saves component status information into configuration file so
it can continue from previous time.

Change-Id: I97801368a0ee0f46aff479249bb48d60fd4a2741
Reviewed-by: Jani Heikkinen <jani.heikkinen@digia.com>