QtWebEngine needs to be separate archive in release packages
Due to licensing terms the QtWebEngine needs to be a separate
installable component in the installer. For this the binary build
script must support creating a separate archive from arbitrary qt5
module.

The installer (install-tree) adaptation needs to be changed in
follow-up commit.

Some related code clean-up and refactoring done as well.

Change-Id: I0c8ddad0cd3407ed2e6e2a889ae2f0d794b8ecd4
Reviewed-by: Antti Kokko <antti.kokko@digia.com>