Add dbus -I and -L flags
This is the output of:
    pkg-config --cflags-only-I dbus-1
    pkg-config --libs-only-L dbus-1

Discussion:

Ideally the configure option we want is -dbus-pkg-config,
i.e. use pkg-config for resolving dbus dependencies
only.

An alternative is to configure with -pkg-config and
-no-* for all dependencies -pgk-config will enable.
This is fragile if the possible dependencies change
and the SDK packaging script is not updated accordingly.

Another alternative is to have the SDK packaging
script run pkg-config.

Change-Id: Ib354dad7950adfd7a4304df0136b336418578261
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Simo Fält <simo.falt@digia.com>
Reviewed-by: Jani Heikkinen <jani.heikkinen@theqtcompany.com>