Remove wrong dependencies to vcredist
qt.tools.vcredist, qt.tools.vcredist_64 contain the runtime
redistributables for Visual Studio 2010 only. It's pointless to let
MSVC2012, MSVC2013 packages depend on it.

(Because Qt Creator is still built with Visual Studio 2010 it will
be automatically included nevertheless. But that might change, as soon
as we start to build Qt Creator with a newer version ...)

Change-Id: I7a6909610a3583f32061c249c884c8ad8b72ce2d
Reviewed-by: Jani Heikkinen <jani.heikkinen@theqtcompany.com>
Reviewed-by: Iikka Eklund <iikka.eklund@theqtcompany.com>