Enable Qt Creator to run on XP
From MSVC 2012 onwards, executables by default do not run anymore on
Windows XP. To work around this, one has to compile against the headers
and libs of an XP compatible platform SDK, and adapt the linker line.

This cheap hack only adapts the linker line, which however seems to be
enough for Qt Creator.

Change-Id: Iae8d822a20959c65589fbd0887cd4817d904c891
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>