Creator: Fix check for running instance
The old logic to get the executable path was broken. Use the path
calculated in by reactOnTargetDirChange instead.

Also remove check for linguist.exe & qmlviewer.exe (which aren't
shiped anymore since ages), and change the working directory of
Qt Creator to the user's home directory, as the standalone installer
does.

Change-Id: Ifcd26894a80556c3ad70f97a215359a05d7929b1
Reviewed-by: Iikka Eklund <iikka.eklund@theqtcompany.com>
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>