Ensure building of deprecated Qt5 components
For Qt5.5 we put some of the qt modules as separately installable
components. Those components will be deprecated in future.

E.g. for qtwebkit this causes issues. The Assistant will get dependency
to qtwebkit and if the user chooses not to install qtwebkit separately
the Assistant will not launch.

We must configure needed Qt packages with '-skip qtwebkit' so that
e.g. the Assistant is not dependent of the qtwebkit. The consequence
is that the top level make will not build qtwebkit anymore. So we need
to explicitly go under those submodules which are marked to be
archived separately and invoke 'make' from there.

Change-Id: I7d996c2600727754445c1c5df03678646c09a7ed
Reviewed-by: Matti Paaso <matti.paaso@theqtcompany.com>