Custom qtwebkit build steps for Qt5.5 packages
QtWebkit is set as separate installable component in the
installers. By default the Assistant would have dependency
into it which we do not want as the qtwebkit installation
is optional now.

As we are configuring Qt5.5 with '-skip qtwebkit' we need
custom build steps to have the qtwebkit libraries built.

- If Qt5.5 build (checked from version number)
- If configure options contain '-skip qtwebkit'

If the above conditions are met then attempt to build
it separately.

Change-Id: I86c05cc285d8c5add928ab310cdc29099b35325f
Reviewed-by: Matti Paaso <matti.paaso@theqtcompany.com>