Fix extra module LD_LIBRARY_PATH handling in RHEL
When building on Redhat with a newer toolchain, we get python from
/opt/rh/python27. bld_app.py makes sure that sys.executable stays in
PATH, but it completely overwrites LD_LIBRARY_PATH, which however is
necessary for the python interpreter. Otherwise subsequent invocations
of python during the build will fail because libpython cannot be found.

As a counter-measure let's try to preserve LD_LIBRARY_PATH and just prepend our
library path.

Change-Id: I2793f9397d6f68ec42f743c8694d235070375f46
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>