Make extra module doc build part optional
Some of the extra/add-on modules are missing documentation.

Previously if the documentation was missing the whole extra module
src pkg creation job failed even if src and examples packages were
created successfully.

Catch the exception if doc package creation fails and allow to
continue uploading src and examples packages to network drive.

Change-Id: Ia4d5b3a7c2d967817f284f4961f4baf2d6a55b37
Reviewed-by: Antti Kokko <antti.kokko@theqtcompany.com>