Fix race condition when downloading and extracting Qt
On Windows, using 7zip in parallel to unzip into the same directory
regularly fails. It checks whether a directory exists and creates it
if not. Creating directories fails if the directory already exists.
So, if a different process (e.g. different 7z) creates the directory
between the check for existence and the creation, the latter fails.

Split the Qt download and extract into a parallel download phase
and a sequential extract phase.

Change-Id: I2e8f841a467267815ad2338b12cc0f388f8e8e57
Reviewed-by: Iikka Eklund <iikka.eklund@theqtcompany.com>