Do not assume that all package paths contain @LICENSE@ by default
Previously the @LICENSE@ was always added into final URL as
we had separate Opensource and Enterprise builds.

Now this has changed as we have unified Qt builds and in addition the
Qt Creator builds will be unified as well.

Do not automatically append the @LICENSE@ into the base path used.

Note that the installer configuration files need to be adjusted
accordingly. So that where needed the @LICENSE@ needs to be added
in package paths.

Change-Id: I9be7221d8f322d6e68624655259690740880a60f
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>
Reviewed-by: Jani Heikkinen <jani.heikkinen@theqtcompany.com>