Allow to assign installer build jobs to dedicated machines
If "assign_to_machine_label:" is defined for the parsed installer job
then check if the current label matches with that. The current machine
label is always present in the "cfg" environment variable.

If the "assign_to_machine_label:" does not exist then check against
the platform and arch as before.

Change-Id: Ie4701e0f7428cd9f2fa1ea1cc5afafb0b290b3ac
Reviewed-by: Matti Paaso <matti.paaso@theqtcompany.com>