Make Qt Creator / Qt 5.6 build possible
For Qt 5.6 we now have one 7zip for each Qt module
instead of groups of modules.
Change the bld_qtcreator (and bld_qtcreator_plugin) script
to take an explicit list of module 7zip URLs instead of
a base URL, so we can make that decision in the build_wrapper
script.
This has the nice side effect that we can handle KDSME
and GammaRay modules like all other Qt modules, even though
they are downloaded from a completely different URL.

Change-Id: I1ad7706e4285d4ad9b1ce168b49775eec5288c87
Reviewed-by: Iikka Eklund <iikka.eklund@theqtcompany.com>