Qt5.6: Rename mingw492 as mingw49 for Qt package configurations
MinGW TC patch level version may change during the Qt5.6 life time so
do not hard code the patch level in the installer configurations of
the corresponding Qt packages.

This makes it also easier for Qt5.6.x online update later on for
MinGW Qt package if the MinGW TC will be updated.

The installer configuration name for MinGW 4.9.2 TC itself will not be
changed as prior minor Qt versions depend on that specific version.

Change-Id: Ib4f5e40c840b80befb18e5144f7aefecebe82910
Reviewed-by: Jani Heikkinen <jani.heikkinen@theqtcompany.com>