Skip qtwebkit when building ifw against Qt5.5.1
Bring back "-skip qtwebkit" as it is still part of Qt5.5.1 and
static compilation of qtwebkit fails (Qt5.5.1) in addition we do not
wan't to include it in the ifw builds.

If/when the installer-framework will be compiled against Qt5.6.x this
part needs some furher work as "-skip qtwebkit" will fail as qtwebkit
is no longer part of sources and the skip option would fail.

Change-Id: I213bdaeba61109a15f2f8e9be327b858a2a14693
Reviewed-by: Katja Marttila <katja.marttila@theqtcompany.com>
Reviewed-by: Antti Kokko <antti.kokko@theqtcompany.com>