build_clang.py: Build with enabled rtti by default
By default, LLVM is build with rtti disabled because they ship their own
implementation of it. However, this does not allow to link LLVM's C++ libs with
other rtti-enabled libs (which is the default), like it is required for Qt
Creator.

Building with enabled rtti comes with a small performance penalty [1]. Because
the C library libclang.dll will be build with mingw [2] and rtti is not
relevant in that regard, let rtti disabled there.

[1] An rtti-enabled libclang.dll build with mingw492 was about 1-3% slower than
    the rtti-disabled version.
[2] The profile guided optimization with mingw yielded better results.

Change-Id: I85cde2f3c932dbd1caf59c837e8cd120503b530c
Reviewed-by: Marco Bubke <marco.bubke@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>