Scripts to generate API change-reviews
Iterator over modules; per-module script to parse arguments and drive
git; API header lister adapted from syncqt.pl's use of sync.profile;
and a commit-filter to revert the boring parts so that the diff is
mostly stuff that actually matters enough to warrant review.  The last
aims to be flexible as we learn more about what's boring.

Task-number: QTQAINFRA-973
Change-Id: Ic4132c9502313ac134a1db89d8f3ef32243fa719
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>