Qt Creator: Install MSVC2015 64bit redistributable with Qt Creator
For 64bit Qt Creator we need the 64bit redist to be able to run
Qt Creator, but for any architecture of Qt Creator we should actually
install both versions of the redistributable if we are on a 64bit system,
since when installing 32bit Qt Creator on a 64bit system, there are
situations like debugging 32bit applications with a 64bit CDB (i.e.
64bit CDB Qt Creator extension).

Add this to 5.6 installers with MSVC2013 based Qt Creator too, to
avoid special casing in the Qt Creator installscript.

Change-Id: I31dce33f5cc58e47181ab7a2a656245b38686a0f
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>