Add job for building Qt Creator plugins
To be used by additional automotive plugin build, and also for 3rdparty
plugins.
Reads a json file with the details of the plugins to be built,
which directly correspond to the QtcPlugin structure, but can
also contain %(VAR)s placeholders which are replaced with the
VAR from optionDict. Modules can be just a Qt module name,
or complete URLs.
Jobs must set QTC_BASE_URL, QT_BASE_URL, and UPLOAD_BASE_DIR,
and platform specific QT_POSTFIX.

Change-Id: I6a5ca71286cdea9e8aa7fb9fbd3996d73208028c
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>