Fix how 'repo_content_type' is parsed from configuration files
The repository builds are put to temporary directories at the
staging server. Now the name of the temporary directory needs to be
defined in the configuration file. The name is usually user friendly
to allow easier 'debugging' if something should fail.

Make it possible to omit the temporary directory name from the
release description files and take it from the used repository build
configuration file.

The temporary directory name will still be unique and user friendly.

Also remove redundant error handling.

Task-number: QTBUG-62349
Change-Id: Ie4ba7714b4fb44aa1172755ba58d0a5e4b4ef677
Reviewed-by: Antti Kokko <antti.kokko@qt.io>