Package the recipe-sequence generation in a generator
The resetboring.Selector.Censor.recipe sequence was set up by code run
in Censor's class body, which made it necessary to del each local
variable used in the process, to avoid polluting Censor's name-space.
It was also getting long (making it worth putting into a function) and
appended to a list (making a generator a natural implementation).

So convert to a generator, called once during the class body, with the
same name as the sequence attribute (that thus replaces it, saving the
need to del it).  It can now have all the local variables it likes
without polluting Censor; and we save the need for the list in which
it was build, now simply gathering the generator's yields into the
tuple we actually want.  Its doc-string also provides a home for some
documentation.

Change-Id: Id1735c38bdbf83d46851f3ff009362ad22a6a329
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>