Qt Creator: Save the sha1(s) that were used for building
Creates a summary file containing the SHA1s for the individual
repositories.
Also create a link from the snapshots base path + job name to
the actually built snapshot which is named by display name / version,
so the build job configurations can find the SHA1s for a specific
job easier.

Change-Id: I591a2c441b50023e90a4b530719a4f79ab660ef7
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>