Don't remove debug symbols from debug info archives on create_installer
Current implementation of REMOVE_DEBUG_INFORMATION_FILES is to check
and remove debug information files from all archives during installer
and repository creation.

Change implementation so that we skip always debug information specific
archives. If surch archives are present in package configurations it
definitely means that debug information is intended to be part of
installer and reposity regardless of removal flag usage.

Change-Id: I3c69f90d4f0a0637f88b004b1d94457078209f46
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>