Qt Creator: Fix missing cross links to Qt documentation
The Qt documentation is in a separate package that does not quite
follow the Qt module naming convention.
Download the documentation package file and repackage in a way that
is compatible with the other Qt modules, so the actual build can
handle the documentation like any other Qt module.

Task-number: QTCREATORBUG-20713
Change-Id: Ifa87cd4a3cc63a4193d3be7fee62692f1afc8682
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>