Fix some indexing issues in resetboring.py
If we're too near to the end of a token list to accommodate a match to
a sequence of words, we can stop searching (else we IndexError).

When replacing one sequence of tokens with another, the search for
more matches for the former is using the indexing in the token list we
started with, before the first replacement. If the replacement's
length doesn't match the original, we need to adjust match positions
to take that into account.

Change-Id: I727576c047be7bfaf047b618c5bc49604e3e7aad
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>