Cope with the new alias for noexcept
We can now have Q_DECL_NOTHROW converted to Q_DECL_NOEXCEPT, or either
of these converted to noexcept. A later commit shall also deal with
Q_DECL_NOEXCEPT_EXPR(), which shall need to happen before this (as it
also turns into noexpr(), albeit with parentheses). Combine with the
other three-step cascade and move both later than it was.

Change-Id: I11b78af9a6eb0948990e4ba50f702c4f93739bb6
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>