Handle Q_DECL_NOEXCEPT_EXPR()'s more common cases
Unfortunately, the way matching and purging is done gets into trouble
if one line contains two matches to a pattern and only one of them has
been changed.  Since noexcept()'s condition often is itself a
noexcept() expression, noexcept(noexcept(...)) happens and we need to
not match it (as the inner noexcept never used the macro). So this
only a partially handle the replacement, but at least it catches the
more common cases.  Added a FIXME comment explaining the problem that
prevents solving this better.

Change-Id: I14b555c67dd9521bdc42adbce82267745ae7a56a
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>