Use environment when checking for virtualenv location
Environment is needed for subprocess because Python may be built with
shared mode thus LD_LIBRARY_PATH is needed.

Also use exact path to used Python installation when invoking the
'pipenv' to avoid issues with system paths containing the 'pipenv'
in it.

Changed to use subprocess.run which accepts the environment as argument.

Change-Id: I3f7bda2d144afa646ff564fc829c6e58766e1896
Reviewed-by: Antti Kokko <antti.kokko@qt.io>