Create a script to manage online repository updates
The script can execute the following tasks:
- build online repositories
- update the (built) online repositories to staging
- trigger RTA for the online repositories
- trigger production sync
  - s3 (aws cli)
  - ext (rsync)

These can be executed all in a row or separately by command
line options.

The input for the script is a top level release description file
which defines all repository tasks including metadata needed
for each step.

For production sync the sync scripts and logs are saved to
the remote server hosting the staging environment:

  <root_dir>/online_repositories/remote_scripts/
  <root_dir>/online_repositories/s3_sync_logs/
  <root_dir>/online_repositories/ext_sync_logs/

If the sync should fail the generated sync scripts are preserved
and can be re-executed if needed.

Tests included.

Change-Id: Ibec57ee32ae7ebb99d7dcb40a4907a5dd7f51611
Reviewed-by: Antti Kokko <antti.kokko@qt.io>