build_clang.py: Build clazy explicitly
The script did not know anything about clazy as clazy was a submodule of
our llvm-project repo, e.g. in the release_80-based branch. Custom
patches in that repository integrated the clazy submodule into the llvm
build.

As we are going to reduce custom patches in our llvm-project and clazy
should not be statically compiled into the clang and libclang artefacts,
we need to adapt the building here.

Change-Id: Ief4215fc540079d208ca2ea9fc447b896621a6b9
Reviewed-by: Cristian Adam <cristian.adam@qt.io>