Simplify online repository updates
There is no need to actually 'update' existing repositories.
We are not doing incremental updates to existing repositories as
the whole repository is always created from scratch.

Also there was the problem with 'updates' that deprecated content
was left in the original repository that was updated.

Copy the new repository in place. If previous one exists move it
as a backup. If previous backup existed then delete it. This means
there will be only one backup available. This saves disk space
and maintenance time.

Task-number: QTQAINFRA-3791
Change-Id: Ic559efd4accdb417d661c69493525ab0872d301b
Reviewed-by: Katja Marttila <katja.marttila@qt.io>
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>