Add script to sign the final installer
On Windows we can just sign the .exe using 'signtool.exe' which is
provisioned by the CI.

On mac we use 'codesign' with hardened runtime option '-o'
to sign the .app bundle so that it can be notarized later.

The final .dmg is then created from the signed .app bundle
using 'hdiutil'.

Task-number: QTIFW-1027
Change-Id: I8b13f3a4df575acf1614c5dff6de9c155757ed89
Reviewed-by: Katja Marttila <katja.marttila@qt.io>