Add support for custom actions scripts per archive in pkg configs
Support "archive_action" per archive in config files. The action
needs to point to existing script file that must be executable.

archive_action: content_cleaner.py, --arg1=abc --arg2=abc

A comma separates the script from arguments.

The '--input-dir, is mandatory option for the action script.
The rest of the options are passed as such to the action script.

A content_cleaner.py script is provided which can be used to clean
content from the extracted archive based on glob rules.

Change-Id: I90e57cdacb1e9140f2b31403c7ef0899b99bf023
Reviewed-by: Antti Kokko <antti.kokko@qt.io>