Qt Creator/macOS: Fix signing of application in disk image
For Qt Creator we provide a disk image with just the application and the
license information, and for that we need to sign the application.
The scripts (in the Qt Creator repository) take that information from
the environment. Since that is gone from the environment in the jobs,
explicitly set it up for the Qt Creator build script.

Change-Id: Iae4ae6eea4c38fe94d8cee2dd50d3f752f4e683a
Reviewed-by: Antti Kokko <antti.kokko@qt.io>