Register packaging events from release_repo_updater.py
Use event_injector.py if --event-injector=<path>
is passed to the script.

This enables sending events from selected places to
database for monitoring purposes.

Use a context manager for START, FINISH and ABORT
events per event name.

This change causes START and FINISH events to be sent
from the following steps:
  - create offline installer(s)
  - create online repositories
  - sync online repositories to production (s3 and ext)

If an exception is raised during the step which causes
event sending then an ABORTED event is sent including
the error message.

Other metadata matching for the task(s) is read from
an export_summary.json file which is optionally available
per integrated product. This metadata it attached into
the events.

Task-number: QTQAINFRA-4442
Change-Id: I3895edde6c7a3c4baa139501dc511167e0f0f97e
Reviewed-by: Antti Kokko <antti.kokko@qt.io>