Fix exception propagation in event_register
The __aexit__ should return False to propagate the exception.
If the --event-injector was not given the class variable
'event_register.event_injector' was not initialized and the return
value was wrongly set.

Fix it so that the return value is set correctly no matter if the
event injector is enabled or not.

Change-Id: Ie1b19deae4befcf39acccedd2af995b6c6aa1da4
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>