Modify script to update package.xml with translations
If new translation ts file is added, update_component_translation.sh
can be run to create the binary (.qm) file of the translation. This
change will add the binary translation file to package.xml if it does
not already exist.

Chinese and Korean translations are skipped because those are not
reviewed yet.

Change-Id: I755775bad96821c9e4ba6031da503aac2232d556
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>
Reviewed-by: Antti Kokko <antti.kokko@qt.io>