QtCreator_libclang: Build Qt Creator training with Qt6
Only build libclang for training. Clang14 fails to build with MinGW GCC 11.2.0
when building extra projects.

We train only with libclang, so build in first step only libclang. The
profile information will be used for the rest of the targets.

Also do not abort on tests failures. Mostly we disable the tests, so better
ignore the failures.

Change-Id: Iaa0226f97ea628aa8e179eab656b8d51f4d0e720
Reviewed-by: Eike Ziller <eike.ziller@qt.io>