Support staging repositories as a source for production updates
Allow using existing staging repositories as the source for production
repository updates.

This reduces the need to run separate production repository builds
if the staging content can be used for distribution.
In this case the staging repository is just copied as the production
repository.

Examples:

  - build repos and update staging
    '--build-repositories --update-staging'
  - build repos and update staging and production
    '--build-repositories --update-staging --update-production'
  - do not build repos but update production from existing staging repos
    '--update-production --update-source=staging'

Introduced a RepoUpdateStrategy class to encapsulate the logic of
different update scenarios there. This class knows what to update
and to where, depending on the use case.

Removed printing of "done" repositories as the output did not make
much sense. The 'handle_update()' returned only the names of the
built repositories, although the 'handle_update()' does also
updates and production sync steps and no verbose output was
given from these. This requires a follow-up patch to fix the
output reporting.

Task-number: QTQAINFRA-5118
Change-Id: If67916035619557e0516eeaca715f44838c87b37
Reviewed-by: Antti Kokko <antti.kokko@qt.io>