Add operations for Debian repo creation and publishing
While the debian_repo_release.py implements a client wrapper
for basic Debian repo operations we need a way to use these
opertions in a meaningful way to actually manage releases.

The first version of release_repo_updater_deb.py implements
a transaction to:
  - create a repository
  - populate the repository with content
  - create snapshot
  - publish the snapshot i.e. make it visible

The transaction is implemented as a command pattern so that
if any of the steps should fail an undo can be performed for
the executed steps in correct order.

Implement a function to crawl URL paths recursively to scan
for .deb package content in a concurrent way.

Implement a function to download given files concurrently
in chunks to minimize memory usage.

Add a test to create and publish a repo with test content
into filesystem endpoint using local Aptly server instance.

Task-number: QTIFW-2813
Change-Id: I6d5bf8ea0c066c35e7c0cfe407c81234327d5c8
Reviewed-by: Antti Kokko <antti.kokko@qt.io>