QtCreator: Sign online/Qt offline bundle on macOS
We need to sign the Qt Creator app that we install with the online and
Qt offline installers, to avoid an issue with access permissions on M1
Macs.

For that the installer may no longer write into the app bundle. The last
remaining thing that the installer wrote, was the optional plugins.
Instead, add a relative plugin path (expanding to "<Qt>/Tools/QtCreator
<version>/PlugIns") to the install settings of Qt Creator, and sign that
bundle. The installer has to install the optional plugins there, instead
of into the bundle. For consistency, do the same with the commercial Qt
Creator offline installers.

Task-number: QTCREATORBUG-26705
Change-Id: I728ebf28b658b23743afb563ccc5cdf96639fa7e
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>