create_installer: Fix substitutions being overridden by common substs
In release_task_reader.py:parse_data(), the common substitutions are
added to the end of the list making it so that they will override any
substitutions set in individual components.

This behavior should be changed so that substitutions from individual
components are preferred over the ones set in common.substitutions.

Disallow the overriding of substitution values in create_installer.py
and log a warning when attempting to do so.

Fixes: QTBUG-109957
Change-Id: I096502fc7f0070efd0819f518267bd7d857f286b
Reviewed-by: Antti Kokko <antti.kokko@qt.io>